refactor: split server bootstrap and wiring
refs #570
